Public Push
This is a public version of a Private Source that I have. It consists of files that I have used to learn Swift.

There are private projects that I created that are a combination of what I learned. If you would like to see them, please contact me with the email you received with my application with your Git username, as there are projects I hope to ship.